refactor: 예외처리 경로 수정
